+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Nuvlabox - Allow Nuvlabox credential to edit reboot action job
   - Nuvlabox - On commission allow Nuvlabox to remove resources
 
 ## [5.11.0] - 2021-03-08